add test file, add cpp version

+++ CompletelyKnapsack_problem/src_c/Knapsack.h
@@ -37,7 +37,7 @@ typedef struct{
     int    Costs1[MaxCaseKinds]; 
     float Costs2[MaxCaseKinds];
     int    limitCost1;
-    int    limitCost2;    
+    float limitCost2;    
 }Knapsack_T;
